check/042 is specific to GFonts onboarding
(issue #1881)
