+++ Lib/fontbakery/specifications/os2.py
@@ -193,26 +193,3 @@ def com_google_fonts_check_034(ttFont):
                   "average of the widths of the latin "
                   "lowercase glyphs in "
                   "the font").format(current_value, expected_value)
-
-
-@check(
-  id = 'com.google.fonts/check/042'
-)
-def com_google_fonts_check_042(ttFont):
-  """Checking OS/2 Metrics match hhea Metrics.
-
-  OS/2 and hhea vertical metric values should match. This will produce
-  the same linespacing on Mac, Linux and Windows.
-
-  Mac OS X uses the hhea values
-  Windows uses OS/2 or Win, depending on the OS or fsSelection bit value.
-  """
-  # OS/2 sTypoAscender and sTypoDescender match hhea ascent and descent
-  if ttFont["OS/2"].sTypoAscender != ttFont["hhea"].ascent:
-    yield FAIL, Message("ascender",
-                        "OS/2 sTypoAscender and hhea ascent must be equal.")
-  elif ttFont["OS/2"].sTypoDescender != ttFont["hhea"].descent:
-    yield FAIL, Message("descender",
-                        "OS/2 sTypoDescender and hhea descent must be equal.")
-  else:
-    yield PASS, ("OS/2.sTypoAscender/Descender" " match hhea.ascent/descent.")
